pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "enhanced-jobbot"
version = "1.0.0"
description = "Automated job application database manager with AI-powered extraction and Notion integration"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Nicole Eid", email = "hello@nicoleeid.me" }
]
keywords = [
    "job-search",
    "automation",
    "notion",
    "ai",
    "web-scraping"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
    "Topic :: Utilities"
]
requires-python = ">=3.8"
dependencies = [
    "openai>=1.0.0",
    "notion-client>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "playwright>=1.40.0",
    "lxml>=4.9.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0"
]

[project.urls]
Homepage = "https://github.com/nicoleeid/enhanced-jobbot"
Repository = "https://github.com/nicoleeid/enhanced-jobbot"
Issues = "https://github.com/nicoleeid/enhanced-jobbot/issues"

[project.scripts]
jobbot = "enhanced_jobbot:main"
jobbot-enhanced = "enhanced_jobbot:main"
jobbot-original = "jobbot:main"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
